# lib/api-zod/src/generated/types/leadUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intessuto Performance Suite API
 * OpenAPI spec version: 0.1.0
 */

export interface LeadUpdate {
  nomeAzienda?: string;
  referente?: string;
  email?: string;
  telefono?: string;
  fonte?: string;
  interesse?: string;
  stato?: string;
  commercialeAssegnato?: string;
  note?: string;
}