# lib/api-zod/src/generated/types/materialeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intessuto Performance Suite API
 * OpenAPI spec version: 0.1.0
 */

export interface MaterialeInput {
  nome: string;
  categoria: string;
  colore: string;
  utilizzoConsigliato: string;
  fasciaPrezzo: string;
  disponibile: boolean;
  descrizione?: string;
}