# lib/api-zod/src/generated/types/proposta.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intessuto Performance Suite API
 * OpenAPI spec version: 0.1.0
 */
import type { PropostaItem } from './propostaItem';
import type { PropostaStato } from './propostaStato';

export interface Proposta {
  id: number;
  clienteId: number;
  clienteNome: string;
  items: PropostaItem[];
  totale: number;
  stato: PropostaStato;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}