# lib/api-zod/src/generated/types/propostaInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intessuto Performance Suite API
 * OpenAPI spec version: 0.1.0
 */
import type { PropostaItem } from './propostaItem';

export interface PropostaInput {
  clienteId: number;
  items: PropostaItem[];
  stato: string;
  note?: string;
}