# lib/api-zod/src/generated/types/richiestaCampioneUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Intessuto Performance Suite API
 * OpenAPI spec version: 0.1.0
 */

export interface RichiestaCampioneUpdate {
  statoSpedizione?: string;
  note?: string;
}